requirements.txt
Django==1.10.5
django-autoslug==1.9.3
mysqlclient==1.3.9
